deleted Team_project.py file and split the program logic into 5 files
